+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Updated
+- badge inclusion and test coverage scripts
+
 ## [0.17.0] - 2021-09-09
 ### Updated
 - DataElements package to 0.4.0